Move all modal components into a folder
